Testing branch for Nix specific changes
Nix doesn't create `/sbin`/`/lib`/`/bin` by default and this breaks the
build process. This is trivial to fix for the `cmd/fakemachine` but if
trying to use the library in `debos` it becomes moree difficult. This is
a patch for my personal flake for `debos` and if it works I'll make a
patchset for fakemachine to add support for specification of bindpaths.